refactor: assets index 경로를 사용하도록 image import 수정 (#224)
