+++ frontend/src/pages/PL/LoadingContainer.tsx
@@ -1,10 +1,7 @@
-import type { DetailedHTMLProps, HTMLAttributes } from 'react'
-import TamnaraIcon from '@/assets/tamnara_icon.png'
+import { TamnaraIcon } from '@/assets'
 import { useTypingLoadingText } from './useTypingLoadingText'
 
-type LoadingContainerProps = DetailedHTMLProps<HTMLAttributes<HTMLDivElement>, HTMLDivElement> & {}
-
-export const LoadingContainer = ({}: LoadingContainerProps) => {
+export const LoadingContainer = () => {
   const iconContainerClass = 'flex flex-col h-full justify-center items-center gap-2'
   const iconClass = 'w-20'
   const textClass = 'text-lg font-semibold text-point'
