+++ frontend/src/pages/PN/PN-003/Thumbnail.tsx
@@ -1,5 +1,5 @@
 import { type DetailedHTMLProps, type HTMLAttributes } from 'react'
-import TamnaraDefaultImage from '@/assets/tamanara_default_image.png'
+import { TamnaraDefaultImg } from '@/assets'
 
 type ReactDivProps = DetailedHTMLProps<HTMLAttributes<HTMLDivElement>, HTMLDivElement>
 type ThumbnailProps = ReactDivProps & {
@@ -10,11 +10,10 @@ type ThumbnailProps = ReactDivProps & {
 export const Thumbnail = ({ image, title }: ThumbnailProps) => {
   const wrapperClass = 'flex overflow-hidden'
   const imageClass = 'w-full h-[200px] object-cover'
-  const fallbackClass = 'w-full h-[200px] bg-carouselBg flex items-center justify-center text-carouselInactive text-sm'
 
   return (
     <div className={wrapperClass}>
-      <img src={image || TamnaraDefaultImage} alt={title} className={imageClass} />
+      <img src={image || TamnaraDefaultImg} alt={title} className={imageClass} />
     </div>
   )
 }
